+++ backend/src/services/PersonsService.ts
@@ -1,8 +1,10 @@
 import { getConnection, getCustomRepository, Repository } from 'typeorm';
-import bcrypt from 'bcrypt';
+// import bcrypt from 'bcrypt';
 import { Person } from '../entities/Person';
 import { PersonsRepository } from '../repositories/PersonsRepository';
 import { Persona } from '../entities/Persona';
+import { OrganizationsRepository } from '../repositories/OrganizationsRepository';
+import { PersonasRepository } from '../repositories/PersonasRepository';
 
 class PersonsService {
     private personsRepository: Repository<Person>;
@@ -13,7 +15,38 @@ class PersonsService {
 
     async create(id_organization: number, id_persona: number, email: string, name: string, password: string, user: string ): Promise<Person> {
         if(!id_organization || !id_persona || !email || !name || !password || !user) {
-            throw new Error('Mandatory values not filled');
+            throw new Error('Campos obrigatórios não preenchidos');
+        }
+
+        if (Number.isNaN(id_organization)) {
+            throw new Error('Organização inválida');
+        }
+
+        if (Number.isNaN(id_persona)) {
+            throw new Error('Persona inválida');
+        }
+
+        // validação de email por regex
+        // eslint-disable-next-line no-useless-escape
+        const regex = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;                          
+        if (!regex.test(email)) {
+            throw new Error('Insira um e-mail valido');
+        }
+
+        const organization = new OrganizationsRepository().findOne({
+            where: { id_organization},
+        });
+
+        if (!organization) {
+            throw new Error('Organização não encontrada');
+        }
+
+        const persona = new PersonasRepository().findOne({
+            where: { id_persona},
+        });
+
+        if (!persona) {
+            throw new Error('Persona não encontrada');
         }
 
         const registeredEmail = await this.personsRepository.findOne({
@@ -21,7 +54,7 @@ class PersonsService {
         });
 
         if (registeredEmail) {
-            throw new Error('E-mail already registered');
+            throw new Error('E-mail já foi registrado');
         }
 
         const registeredPerson = await this.personsRepository.findOne({
@@ -29,7 +62,7 @@ class PersonsService {
         });
 
         if (registeredPerson) {
-            throw new Error('Username already registered');
+            throw new Error('Usuário já foi registrado');
         }
 
         // password = await bcrypt.hash(password, 10);
@@ -51,22 +84,53 @@ class PersonsService {
     }
 
     async list(id_organization: number): Promise<Person[]> {
+
+        if (!id_organization) {
+            throw new Error('Campos obrigatórios não preenchidos');
+        }
+
+        if (Number.isNaN(id_organization)) {
+            throw new Error('Organização inválida');
+        }
+
+        const organization = new OrganizationsRepository().findOne({
+            where: { id_organization},
+        });
+
+        if (!organization) {
+            throw new Error('Organização não encontrada');
+        }
+
         const list = await getConnection()
-                            .createQueryBuilder(Person, 'person')
-                            .select('person.id_person', 'id_person')
-                            .addSelect('person.id_persona', 'id_persona')
-                            .addSelect('person.id_organization', 'id_organization')
-                            .addSelect('person.name', 'name')
-                            .addSelect('person.user', 'user')
-                            .addSelect('person.email', 'email')
-                            .addSelect('persona.type_persona', 'type_persona')
-                            .leftJoin(Persona, 'persona', 'person.id_persona = persona.id_persona')
-                            .where('person.id_organization = :id_organization', { id_organization})
-                            .getRawMany();
+        .createQueryBuilder(Person, 'person')
+        .select('person.id_person', 'id_person')
+        .addSelect('person.id_persona', 'id_persona')
+        .addSelect('person.id_organization', 'id_organization')
+        .addSelect('person.name', 'name')
+        .addSelect('person.user', 'user')
+        .addSelect('person.email', 'email')
+        .addSelect('persona.type_persona', 'type_persona')
+        .leftJoin(Persona, 'persona', 'person.id_persona = persona.id_persona')
+        .where('person.id_organization = :id_organization', { id_organization})
+        .getRawMany();
+
+        if (list.length < 1) {
+            throw new Error('Nenhuma pessoa está cadastrada');
+        }
+
         return list;
     }
 
     async findById(id_person: number): Promise<Person> {
+
+        if (!id_person) {
+            throw new Error('Campos obrigatórios não preenchidos');
+        }
+
+        if (Number.isNaN(id_person)) {
+            throw new Error('Pessoa inválida');
+        }
+
         const person = await getConnection()
         .createQueryBuilder(Person, 'person')
         .select('person.id_person', 'id_person')
@@ -85,12 +149,51 @@ class PersonsService {
 
     async updateById(id_person: number, id_organization: number, id_persona: number, email: string, name: string, user: string, oldPassword: string, password?: string): Promise<Person>{
 
+        if(!id_person || !id_organization || !id_persona || !email || !name || !oldPassword || !user) {
+            throw new Error('Campos obrigatórios não preenchidos');
+        }
+
+        if (Number.isNaN(id_person)) {
+            throw new Error('Pessoa inválida');
+        }
+
+        if (Number.isNaN(id_organization)) {
+            throw new Error('Organização inválida');
+        }
+
+        if (Number.isNaN(id_persona)) {
+            throw new Error('Persona inválida');
+        }
+
+        // validação de email por regex
+        // eslint-disable-next-line no-useless-escape
+        const regex = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;                          
+        if (!regex.test(email)) {
+            throw new Error('Insira um e-mail valido');
+        }
+
+        const organization = new OrganizationsRepository().findOne({
+            where: { id_organization},
+        });
+
+        if (!organization) {
+            throw new Error('Organização não encontrada');
+        }
+
+        const persona = new PersonasRepository().findOne({
+            where: { id_persona},
+        });
+
+        if (!persona) {
+            throw new Error('Persona não encontrada');
+        }
+
         const registeredEmail = await this.personsRepository.findOne({
             where: {email},
         });
 
         if (registeredEmail && registeredEmail.id_person !== id_person) {
-            throw new Error('E-mail already registered');
+            throw new Error('E-mail já existe');
         }
 
         const registeredPerson = await this.personsRepository.findOne({
@@ -98,7 +201,7 @@ class PersonsService {
         });
 
         if (registeredPerson && registeredPerson.id_person !== id_person) {
-            throw new Error('Username already registered');
+            throw new Error('Usuário já existe');
         }
 
         const person = await this.personsRepository.findOne({
@@ -106,11 +209,11 @@ class PersonsService {
         });
 
         if (person.password !== oldPassword) {
-            throw new Error('Password doesn\'t match');
+            throw new Error('Senha incorreta');
         }
 
         if (oldPassword === password) {
-            throw new Error('Insert a new password');
+            throw new Error('Insira uma nova senha');
         }
 
         person.name = name;
@@ -128,12 +231,21 @@ class PersonsService {
     }
 
     async deleteById(id_person: number): Promise<boolean> {
+
+        if (!id_person) {
+            throw new Error('Campos obrigatórios não preenchidos');
+        }
+
+        if (Number.isNaN(id_person)) {
+            throw new Error('Pessoa inválida');
+        }
+
         const person = await this.personsRepository.findOne({
             where: {id_person},
         });
 
         if (!person) {
-            throw new Error('Person doesn\'t exist');
+            throw new Error('Pessoa não existe');
         }
 
         await this.personsRepository.delete(id_person);
